second commit

+++ src/Point.java
@@ -1,13 +1,13 @@
 public class Point {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
     public Point(){
         this.x=0;
         this.y=0;
     }
 
-    public Point(int x,int y){
+    public Point(float x,float y){
         this.x=x;
         this.y=y;
     }
@@ -16,22 +16,21 @@ public class Point {
         this.y = other.y;
     }
 
-    public void set(int x,int y){
+    public void set(float x,float y){
         this.x=x;
         this.y=y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
-    public int getY(){
+    public float getY(){
         return y;
     }
 
     public Point add(Point other) {
         return new Point(this.x + other.x, this.y + other.y);
     }
-
 
     @Override
     public String toString() {
